requirements.txt
pypinyin
absl-py